daft-quickstart/requirements.txt
getdaft[all]
psycopg2-binary
apache-beam[test]
litecli
pyspark
polars